More consistent playlist responses
This makes the /now endpoint and the /playlists endpoint reply
with similar objects as /playlists/:id/. Particularly, they now
include the "size" property that is present on /playlists/:id, and
they do _not_ include the media ID array that they had before.

This also future-proofs the /now endpoint so that the responses it
gives will always be of the same shape as the /booth and /playlists
endpoints, so both the server and clients can reuse that code.