playlists: add pagination metadata to /media requests
This splits the getPlaylist controller function into two simpler
functions: getPlaylist and getPlaylistItems. (The getPlaylist
function was essentially already used for two entirely different
things and controlled by a boolean flag `populate`.)

This also removes the pagination from a normal `getPlaylist` call,
because there is no longer anything to paginate (the `media` array
is replaced by a `size` number in `toPlaylistResponse`).

Getting playlist _items_ is moved into getPlaylistItems, which does
have pagination, and also returns a paginated object, which wasn't
possible before because the items would live on the `.media`
property of a playlist and paginating a property is weird.