*: async functions
Builds on #37.

So this is pretty massive, but it's mostly applying syntax sugar.
This makes a bunch of stuff much easier to follow, by flattening
deep promise trees, and removing lots of temporary variables that
are no longer necessary.

There are a few behavioural changes too: a few more `await`s have
been added around the place, mostly around Redis calls. I don't
know if all of them are necessary, but at least some of them were,
because this fixes a good chunk of the waitlist sync issues that
we had before.