advance: refactor advances, fixes waitlist sync
This pulls most of the advance logic into ./advance and into a
single internal "advance" command, instead of the previous
"cycleWaitlist" and "advance" commands. Previously, the current DJ
would be added to the waitlist again, which would trigger a
waitlist update, and an "advance" command. The "advance" command
would take the first DJ from the waitlist, process the advance,
and broadcast a public "advance" command. No new waitlist update
would be triggered, so clients' waitlists would get out of sync.

This patch changes the advance() function to do the waitlist
updates, the history updates _and_ the booth updates. So now when
you call the advance() function, you can just grab the updated
waitlist at the end and it'll be consistent (save for errors...)