sockets: rewrite SocketServer to use Connection types
Introduces connection classes:

  - GuestConnection: an unauthenticated client
  - AuthedConnection: an authenticated user
  - LostConnection: a lost connection to an authenticated user,
    used to reconnect them properly when they return. (After a
    refresh, for example.)

Messages to LostConnections are stored in Redis so reconnects work
properly when the server restarts.

When an authenticated user disconnects, their AuthedConnection
is replaced by a LostConnection, which waits for 30 seconds for a
reconnect by the user before notifying the server that the user
left the room. During that 30 second window, other people can
still @-mention the disconnected user, and the user will stay in
the waitlist or booth. That deals with most cases of accidental
refresh etc., and 30 seconds is usually also more than enough for
server restarts.

When the server starts up, LostConnections are created for all
users that were previously on the server, so they'll have 30
seconds to reconnect.