api: add "server" option, replacing obsolete üWave core server
This turns the ApiV1 class into an Actual Router™, instead of a
class wrapping around a router.

A "server" or "socketPort" option now needs to be passed in to
attach the WebSocket server to the right server/port.

Some of the exceptions in the authentication middleware were
changed to silent returns to make sure that people with invalid
tokens can still access public routes.